project/backend/job_parser/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
django-cron==0.6.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-rest-elasticsearch==0.4.2
djangorestframework==3.15.2
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
idna==3.7
kombu==5.3.7
packaging==24.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
